PR-03: extract config.rs and focus.rs from main.rs
Move Rule, NativeTerminalRule, ConfigEntry, Config, WindowInfo,
load_config, match_pattern into config.rs; move FocusAction,
FocusActions, FocusHandler (and NATIVE_TERMINAL_RULE_INDEX) into
focus.rs. All items made pub(crate). Add mod/use/re-export
declarations to main.rs. Remove regex::Regex import from main.rs
(made unused by the move). 261 tests pass.
