+++ src/daemon/main.rs
@@ -5,7 +5,6 @@ use ksni::{Icon as SniIcon, MenuItem, Status as SniStatus, ToolTip, Tray, TraySe
 use noto_sans_mono_bitmap::{
     FontWeight, RasterHeight, RasterizedChar, get_raster, get_raster_width,
 };
-use regex::Regex;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::env;
@@ -92,15 +91,19 @@ mod constants;
 mod errors;
 mod environ;
 mod dbus_naming;
+mod config;
+mod focus;
 
 use constants::*;
 use errors::DynError;
 use environ::*;
 use dbus_naming::*;
+use config::*;
+use focus::*;
 
 #[cfg(test)]
 #[allow(unused_imports)]
-pub(crate) use crate::{constants::*, errors::*, environ::*, dbus_naming::*};
+pub(crate) use crate::{constants::*, errors::*, environ::*, dbus_naming::*, config::*, focus::*};
 
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
@@ -563,559 +566,6 @@ async fn send_control_command_broadcast(
     Ok(BroadcastReport { results })
 }
 
-// === Config ===
-
-/// A rule for matching windows and triggering actions.
-/// At least one of `layer`, `virtual_key`, or `raw_vk_action` should be specified.
-#[derive(Debug, Clone, Deserialize)]
-#[serde(deny_unknown_fields)]
-struct Rule {
-    class: Option<String>,
-    title: Option<String>,
-    /// Layer to switch to when switching to a native terminal (VT)
-    on_native_terminal: Option<String>,
-    /// Layer to switch to when rule matches
-    layer: Option<String>,
-    /// Virtual key to press while window is focused (auto-released on unfocus)
-    virtual_key: Option<String>,
-    /// Raw virtual key actions to fire on focus (fire-and-forget)
-    /// Format: [["vk_name", "Press|Release|Tap|Toggle"], ...]
-    raw_vk_action: Option<Vec<(String, String)>>,
-    /// Continue matching subsequent rules after this one
-    #[serde(default)]
-    fallthrough: bool,
-}
-
-#[derive(Debug, Clone)]
-struct NativeTerminalRule {
-    layer: String,
-    virtual_key: Option<String>,
-    raw_vk_action: Vec<(String, String)>,
-}
-
-#[derive(Debug, Clone)]
-enum ConfigEntry {
-    Default { default: String },
-    Rule(Rule),
-}
-
-impl<'de> serde::Deserialize<'de> for ConfigEntry {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        use serde::de::Error;
-
-        let value = serde_json::Value::deserialize(deserializer)?;
-
-        // Check if it's a "default" entry
-        if let Some(obj) = value.as_object() {
-            if obj.contains_key("default") {
-                if obj.len() == 1 {
-                    if let Some(default) = obj.get("default").and_then(|v| v.as_str()) {
-                        return Ok(ConfigEntry::Default {
-                            default: default.to_string(),
-                        });
-                    }
-                }
-                return Err(D::Error::custom(
-                    "'default' entry should only contain the 'default' field",
-                ));
-            }
-        }
-
-        // Try to parse as Rule with custom error handling for unknown fields
-        let known_fields = [
-            "class",
-            "title",
-            "on_native_terminal",
-            "layer",
-            "virtual_key",
-            "raw_vk_action",
-            "fallthrough",
-        ];
-
-        if let Some(obj) = value.as_object() {
-            for key in obj.keys() {
-                if !known_fields.contains(&key.as_str()) {
-                    return Err(D::Error::custom(format!(
-                        "unknown field '{}'. Valid fields are: class, title, on_native_terminal, layer, virtual_key, raw_vk_action, fallthrough",
-                        key
-                    )));
-                }
-            }
-        }
-
-        serde_json::from_value(value)
-            .map(ConfigEntry::Rule)
-            .map_err(D::Error::custom)
-    }
-}
-
-#[derive(Debug, Clone)]
-struct Config {
-    rules: Vec<Rule>,
-    default_layer: Option<String>,
-    native_terminal_rule: Option<NativeTerminalRule>,
-}
-
-#[derive(Debug, Clone, Default, Deserialize)]
-struct WindowInfo {
-    class: String,
-    title: String,
-    #[serde(default)]
-    is_native_terminal: bool,
-}
-
-fn load_config(config_path: Option<&Path>) -> Config {
-    let path = config_path.map(|p| p.to_path_buf()).unwrap_or_else(|| {
-        let xdg_config = env::var("XDG_CONFIG_HOME")
-            .map(PathBuf::from)
-            .unwrap_or_else(|_| dirs::home_dir().unwrap().join(".config"));
-        xdg_config.join("kanata").join("kanata-switcher.json")
-    });
-
-    if !path.exists() {
-        eprintln!("[Config] Error: Config file not found: {}", path.display());
-        eprintln!();
-        eprintln!("Example config:");
-        eprintln!(
-            r#"[
-  {{"default": "base"}},
-  {{"on_native_terminal": "tty"}},
-  {{"class": "firefox", "layer": "browser"}},
-  {{"class": "alacritty", "title": "vim", "layer": "vim"}}
-]"#
-        );
-        std::process::exit(1);
-    }
-
-    match fs::read_to_string(&path) {
-        Ok(content) => match serde_json::from_str::<Vec<ConfigEntry>>(&content) {
-            Ok(entries) => {
-                let mut rules = Vec::new();
-                let mut default_layer: Option<String> = None;
-                let mut native_terminal_rule: Option<NativeTerminalRule> = None;
-
-                for entry in entries {
-                    match entry {
-                        ConfigEntry::Default { default } => {
-                            if default_layer.is_some() {
-                                eprintln!(
-                                    "[Config] Error: multiple 'default' entries found, only one allowed"
-                                );
-                                std::process::exit(1);
-                            }
-                            default_layer = Some(default);
-                        }
-                        ConfigEntry::Rule(rule) => {
-                            if let Some(layer) = rule.on_native_terminal.clone() {
-                                if rule.class.is_some() || rule.title.is_some() {
-                                    eprintln!(
-                                        "[Config] Error: 'on_native_terminal' cannot be combined with 'class' or 'title'"
-                                    );
-                                    std::process::exit(1);
-                                }
-                                if rule.layer.is_some() {
-                                    eprintln!(
-                                        "[Config] Error: 'on_native_terminal' cannot be combined with 'layer'"
-                                    );
-                                    std::process::exit(1);
-                                }
-                                if native_terminal_rule.is_some() {
-                                    eprintln!(
-                                        "[Config] Error: multiple 'on_native_terminal' rules found, only one allowed"
-                                    );
-                                    std::process::exit(1);
-                                }
-                                native_terminal_rule = Some(NativeTerminalRule {
-                                    layer,
-                                    virtual_key: rule.virtual_key.clone(),
-                                    raw_vk_action: rule.raw_vk_action.clone().unwrap_or_default(),
-                                });
-                            } else {
-                                // Rule with no matchers and no fallthrough would match everything
-                                // and stop further matching, which is almost certainly a bug
-                                if rule.class.is_none() && rule.title.is_none() && !rule.fallthrough
-                                {
-                                    eprintln!(
-                                        "[Config] Error: Rule with no 'class' or 'title' matcher requires 'fallthrough: true'"
-                                    );
-                                    eprintln!(
-                                        "[Config] Hint: A catch-all rule without fallthrough would match all windows and stop further matching"
-                                    );
-                                    std::process::exit(1);
-                                }
-                                rules.push(rule);
-                            }
-                        }
-                    }
-                }
-
-                println!(
-                    "[Config] Loaded {} rules from {}",
-                    rules.len(),
-                    path.display()
-                );
-
-                Config {
-                    rules,
-                    default_layer,
-                    native_terminal_rule,
-                }
-            }
-            Err(e) => {
-                eprintln!("[Config] Error: Failed to parse {}: {}", path.display(), e);
-                std::process::exit(1);
-            }
-        },
-        Err(e) => {
-            eprintln!("[Config] Error: Failed to read {}: {}", path.display(), e);
-            std::process::exit(1);
-        }
-    }
-}
-
-fn match_pattern(pattern: Option<&str>, value: &str) -> bool {
-    match pattern {
-        None => true,
-        Some("*") => true,
-        Some(pat) => match Regex::new(pat) {
-            Ok(re) => re.is_match(value),
-            Err(_) => value.contains(pat),
-        },
-    }
-}
-
-// === Focus Handler ===
-
-/// Individual action to execute on focus change
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum FocusAction {
-    /// Release a virtual key
-    ReleaseVk(String),
-    /// Switch to a layer
-    ChangeLayer(String),
-    /// Press and hold a virtual key (managed - will be released on next focus change)
-    PressVk(String),
-    /// Raw VK action (name, action: Press/Release/Tap/Toggle)
-    RawVkAction(String, String),
-}
-
-/// Actions to execute on focus change, in order.
-/// With fallthrough, all matching actions are collected and executed sequentially.
-#[derive(Debug, Default, Clone, PartialEq, Eq)]
-struct FocusActions {
-    /// Ordered list of actions to execute
-    actions: Vec<FocusAction>,
-    /// The new ordered list of managed VKs after execution (pressed top-to-bottom, released bottom-to-top)
-    new_managed_vks: Vec<String>,
-}
-
-impl FocusActions {
-    fn is_empty(&self) -> bool {
-        self.actions.is_empty()
-    }
-}
-
-const NATIVE_TERMINAL_RULE_INDEX: usize = usize::MAX;
-
-#[derive(Debug)]
-struct FocusHandler {
-    rules: Vec<Rule>,
-    native_terminal_rule: Option<NativeTerminalRule>,
-    last_class: String,
-    last_title: String,
-    last_matched_rules: Vec<usize>,
-    last_effective_layer: String,
-    /// Currently held virtual keys, in order they were pressed (top-to-bottom rule order)
-    current_virtual_keys: Vec<String>,
-    quiet_focus: bool,
-}
-
-impl FocusHandler {
-    fn new(
-        rules: Vec<Rule>,
-        native_terminal_rule: Option<NativeTerminalRule>,
-        quiet_focus: bool,
-    ) -> Self {
-        Self {
-            rules,
-            native_terminal_rule,
-            last_class: String::new(),
-            last_title: String::new(),
-            last_matched_rules: Vec::new(),
-            last_effective_layer: String::new(),
-            current_virtual_keys: Vec::new(),
-            quiet_focus,
-        }
-    }
-
-    /// Handle a focus change event. Returns actions to execute.
-    /// With fallthrough, ALL matching actions are collected and executed in order.
-    /// All matched virtual_keys are pressed and held simultaneously.
-    fn handle(&mut self, win: &WindowInfo, default_layer: &str) -> Option<FocusActions> {
-        let mut result = FocusActions::default();
-
-        if win.is_native_terminal {
-            return self.handle_native_terminal(default_layer);
-        }
-
-        // Handle unfocused state (no window has focus)
-        if win.class.is_empty() && win.title.is_empty() {
-            return self.handle_unfocused(default_layer);
-        }
-
-        if !self.quiet_focus {
-            println!("[Focus] class=\"{}\" title=\"{}\"", win.class, win.title);
-        }
-
-        // Match rules with fallthrough support
-        struct MatchedRule {
-            index: usize,
-            layer: Option<String>,
-            virtual_key: Option<String>,
-            raw_vk_actions: Vec<(String, String)>,
-        }
-
-        let mut matched_rules: Vec<MatchedRule> = Vec::new();
-
-        for (index, rule) in self.rules.iter().enumerate() {
-            if match_pattern(rule.class.as_deref(), &win.class)
-                && match_pattern(rule.title.as_deref(), &win.title)
-            {
-                matched_rules.push(MatchedRule {
-                    index,
-                    layer: rule.layer.clone(),
-                    virtual_key: rule.virtual_key.clone(),
-                    raw_vk_actions: rule.raw_vk_action.clone().unwrap_or_default(),
-                });
-
-                if !rule.fallthrough {
-                    break;
-                }
-            }
-        }
-
-        let matched_indices: Vec<usize> = matched_rules.iter().map(|rule| rule.index).collect();
-
-        // Collect all VKs from matched rules in order (for holding)
-        let new_vks: Vec<String> = matched_rules
-            .iter()
-            .filter_map(|r| r.virtual_key.clone())
-            .collect();
-
-        // Release VKs that are no longer matched (in reverse order)
-        for vk in self.current_virtual_keys.iter().rev() {
-            if !new_vks.contains(vk) {
-                result.actions.push(FocusAction::ReleaseVk(vk.clone()));
-            }
-        }
-
-        // If no rules matched, use default layer
-        if matched_rules.is_empty() {
-            if !default_layer.is_empty() && self.last_effective_layer != default_layer {
-                result
-                    .actions
-                    .push(FocusAction::ChangeLayer(default_layer.to_string()));
-            }
-            result.new_managed_vks = Vec::new();
-            self.last_effective_layer = default_layer.to_string();
-        } else {
-            let matched_changed = matched_indices != self.last_matched_rules;
-            let mut matched_layers: Vec<String> = Vec::new();
-            for matched in &matched_rules {
-                if let Some(layer) = matched.layer.clone() {
-                    matched_layers.push(layer);
-                }
-            }
-            let new_rules: Vec<usize> = matched_indices
-                .iter()
-                .cloned()
-                .filter(|idx| !self.last_matched_rules.contains(idx))
-                .collect();
-
-            // Process matched rules in order, building action list
-            for matched in matched_rules {
-                let is_new = new_rules.contains(&matched.index);
-                if is_new {
-                    // Layer change
-                    if let Some(layer) = matched.layer {
-                        result.actions.push(FocusAction::ChangeLayer(layer));
-                    }
-
-                    // Virtual key: press if not already held
-                    if let Some(ref vk) = matched.virtual_key {
-                        if !self.current_virtual_keys.contains(vk) {
-                            result.actions.push(FocusAction::PressVk(vk.clone()));
-                        }
-                    }
-
-                    // Raw VK actions
-                    for (name, action) in matched.raw_vk_actions {
-                        result.actions.push(FocusAction::RawVkAction(name, action));
-                    }
-                }
-            }
-
-            if matched_changed {
-                if let Some(new_layer) = matched_layers.last().cloned() {
-                    if self.last_effective_layer != new_layer {
-                        let has_new_layer = result.actions.iter().rev().find_map(|action| {
-                            if let FocusAction::ChangeLayer(layer) = action {
-                                Some(layer == &new_layer)
-                            } else {
-                                None
-                            }
-                        });
-                        if has_new_layer != Some(true) {
-                            result
-                                .actions
-                                .push(FocusAction::ChangeLayer(new_layer.clone()));
-                        }
-                    }
-                    self.last_effective_layer = new_layer;
-                }
-            }
-
-            result.new_managed_vks = new_vks;
-        }
-
-        // Update state
-        self.last_class = win.class.clone();
-        self.last_title = win.title.clone();
-        self.last_matched_rules = matched_indices;
-        self.current_virtual_keys = result.new_managed_vks.clone();
-
-        if result.is_empty() {
-            None
-        } else {
-            Some(result)
-        }
-    }
-
-    fn current_virtual_keys(&self) -> Vec<String> {
-        self.current_virtual_keys.clone()
-    }
-
-    fn reset(&mut self) {
-        self.last_class.clear();
-        self.last_title.clear();
-        self.last_matched_rules.clear();
-        self.last_effective_layer.clear();
-        self.current_virtual_keys.clear();
-    }
-
-    fn handle_unfocused(&mut self, default_layer: &str) -> Option<FocusActions> {
-        let mut result = FocusActions::default();
-        if !self.quiet_focus {
-            println!("[Focus] No window focused");
-        }
-        // Release all active virtual keys in reverse order (bottom-to-top)
-        for vk in self.current_virtual_keys.iter().rev() {
-            result.actions.push(FocusAction::ReleaseVk(vk.clone()));
-        }
-        // Switch to default layer
-        if !default_layer.is_empty() && self.last_effective_layer != default_layer {
-            result
-                .actions
-                .push(FocusAction::ChangeLayer(default_layer.to_string()));
-        }
-        result.new_managed_vks = Vec::new();
-        self.current_virtual_keys = Vec::new();
-        self.last_matched_rules.clear();
-        self.last_effective_layer = default_layer.to_string();
-        self.last_class.clear();
-        self.last_title.clear();
-        if result.is_empty() {
-            None
-        } else {
-            Some(result)
-        }
-    }
-
-    fn handle_native_terminal(&mut self, default_layer: &str) -> Option<FocusActions> {
-        let Some(rule) = self.native_terminal_rule.clone() else {
-            return self.handle_native_terminal_without_rule(default_layer);
-        };
-
-        if !self.quiet_focus {
-            println!("[Focus] Native terminal active");
-        }
-
-        let mut result = FocusActions::default();
-        let mut new_vks = Vec::new();
-
-        if let Some(vk) = rule.virtual_key.clone() {
-            new_vks.push(vk);
-        }
-
-        for vk in self.current_virtual_keys.iter().rev() {
-            if !new_vks.contains(vk) {
-                result.actions.push(FocusAction::ReleaseVk(vk.clone()));
-            }
-        }
-
-        let matched_indices = vec![NATIVE_TERMINAL_RULE_INDEX];
-        let is_new = self.last_matched_rules != matched_indices;
-
-        if is_new {
-            if !rule.layer.is_empty() && self.last_effective_layer != rule.layer {
-                result
-                    .actions
-                    .push(FocusAction::ChangeLayer(rule.layer.clone()));
-            }
-            if let Some(vk) = rule.virtual_key {
-                if !self.current_virtual_keys.contains(&vk) {
-                    result.actions.push(FocusAction::PressVk(vk));
-                }
-            }
-            for (name, action) in rule.raw_vk_action {
-                result.actions.push(FocusAction::RawVkAction(name, action));
-            }
-        }
-
-        result.new_managed_vks = new_vks;
-        self.last_matched_rules = matched_indices;
-        self.last_effective_layer = rule.layer;
-        self.current_virtual_keys = result.new_managed_vks.clone();
-        self.last_class.clear();
-        self.last_title.clear();
-
-        if result.is_empty() {
-            None
-        } else {
-            Some(result)
-        }
-    }
-
-    fn handle_native_terminal_without_rule(&mut self, default_layer: &str) -> Option<FocusActions> {
-        let mut result = FocusActions::default();
-        if !self.quiet_focus {
-            println!("[Focus] Native terminal active");
-        }
-        for vk in self.current_virtual_keys.iter().rev() {
-            result.actions.push(FocusAction::ReleaseVk(vk.clone()));
-        }
-        if !default_layer.is_empty() && self.last_effective_layer != default_layer {
-            result
-                .actions
-                .push(FocusAction::ChangeLayer(default_layer.to_string()));
-        }
-        result.new_managed_vks = Vec::new();
-        self.current_virtual_keys = Vec::new();
-        self.last_matched_rules.clear();
-        self.last_effective_layer = default_layer.to_string();
-        self.last_class.clear();
-        self.last_title.clear();
-
-        if result.is_empty() {
-            None
-        } else {
-            Some(result)
-        }
-    }
-}
 
 #[derive(Clone, Debug, PartialEq, Eq)]
 struct StatusSnapshot {
